# Remove ExecutionNode.IsResultSet and memory for Source