--- a/src/GraphQL.ApiTests/GraphQL.SystemReactive.approved.txt
+++ b/src/GraphQL.ApiTests/GraphQL.SystemReactive.approved.txt
@@ -4,6 +4,7 @@ namespace GraphQL.Execution
     {
         public SubscriptionExecutionStrategy() { }
         public static GraphQL.Execution.SubscriptionExecutionStrategy Instance { get; }
+        protected GraphQL.Execution.ExecutionNode BuildSubscriptionExecutionNode(GraphQL.Execution.ExecutionNode parent, GraphQL.Types.IGraphType graphType, GraphQL.Language.AST.Field field, GraphQL.Types.FieldType fieldDefinition, int? indexInParentNode, object source) { }
         public override System.Threading.Tasks.Task<GraphQL.ExecutionResult> ExecuteAsync(GraphQL.Execution.ExecutionContext context) { }
         protected virtual System.Threading.Tasks.Task<System.IObservable<GraphQL.ExecutionResult>> ResolveEventStreamAsync(GraphQL.Execution.ExecutionContext context, GraphQL.Execution.ExecutionNode node) { }
     }