Should provide 16 bytes of savings for every instance of ExecutionNode, including ValueExecutionNode instances.  Have not yet run benchmarks

### Concept

* `Source` is only set during subscription execution.  It is left `null` for any other execution.  To reduce memory, `_source` has been removed from `ExecutionNode` along with the setter, and for subscriptions, another class is used that overrides the source property.
* `IsResultSet` is unnecessary, since the execution strategy almost always is required to process nodes like this: 1. create node 2. execute node 3. execute data loader, if any.   The only exception is that root nodes are not executed, and children of array nodes are not executed.  A simple type check (`node is RootExecutionNode` or `node.Parent is ArrayExecutionNode`) can determine this and skip execution.  Step 3 (execute data loader, if any) does not use `IsResultSet` but simply performs a type check on the result (`result is IDataLoaderResult`).  So, `IsResultSet` is unnecessary.  In order to verify this concept, I added `if (!IsResultSet) throw new Exception();` code to all the locations where `IsResultSet` was accessed, ran all the tests, and none of the exceptions were triggered.
* These two properties consumes 9 bytes on a 64-bit machine, and probably saves 16 bytes due to alignment.
* As `ValueExecutionNode` is one of the most memory-consuming objects overall during execution (per #2237), this represents a significant savings during execution.