@Shane32 2021-01-31
The other stages were unchanged.  Seems about 1% slower and 3-7% faster

### before

|        Method |           Stage |           Mean |         Error |   Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |---------------- |---------------:|--------------:|--------:|--------:|------:|----------:|
| Introspection |         Execute | 660,155.618 ns | 4,428.2779 ns | 50.7813 | 15.6250 |     - |  322306 B |
|          Hero |         Execute |   2,165.165 ns |     9.6959 ns |  0.3586 |       - |     - |    2264 B |
|     Variables |         Execute |   2,811.039 ns |    12.9771 ns |  0.3204 |       - |     - |    2024 B |
|       Literal |         Execute |  29,800.789 ns |   169.0502 ns |  2.0752 |       - |     - |   13127 B |

### after

|        Method |           Stage |           Mean |         Error |   Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |---------------- |---------------:|--------------:|--------:|--------:|------:|----------:|
| Introspection |         Execute | 669,300.467 ns | 3,322.1543 ns | 47.8516 | 14.6484 |     - |  300314 B |
|          Hero |         Execute |   2,196.080 ns |     5.2176 ns |  0.3471 |       - |     - |    2200 B |
|     Variables |         Execute |   2,737.731 ns |    10.0248 ns |  0.3166 |       - |     - |    1992 B |
|       Literal |         Execute |  29,981.630 ns |    74.8146 ns |  2.0752 |       - |     - |   13095 B |