@Shane32 2021-01-31
> I'm finished review. Are you going to do benchmarks again after all the changes from updated `develop`?

I hadn't planned to.  These changes have nothing to do with the other changes.